backend/simple/notes/requirements.txt
grpcio
grpcio-tools
boto3
mysql-connector-python
python-dotenv
PyMuPDF
SQLAlchemy
psycopg2-binary